refactor(http-server): split rmcp_tool_call_dispatch for file-size limit
Replace the 1597-line monolith with a module tree (wire, thread_route,
helpers, handlers/*) so each file stays under the 1500-line CI cap.
Public API unchanged: dispatch_rmcp_tool_call, dispatch_action_with_thread_routing,
decode_dispatch_output, use_main_thread_route, call_meta_from_rmcp.
