+++ crates/dcc-mcp-http-server/src/rmcp_tool_call_dispatch/handlers/mod.rs
@@ -1,0 +1,9 @@
+mod discovery;
+mod dynamic;
+mod jobs;
+mod skills;
+
+pub(super) use discovery::*;
+pub(super) use dynamic::*;
+pub(super) use jobs::*;
+pub(super) use skills::*;
